requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
httpx==0.24.1
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
brotli 